+++ docs/installation_configuration_guide/ModelConfiguration.md
@@ -29,7 +29,6 @@ Look at [configuration areas](./Configuration#configuration-areas) to understand
 | enableObjectReuse                                  | Low        | boolean  | true                   | Should allow [object reuse](https://ci.apache.org/projects/flink/flink-docs-stable/docs/dev/execution/execution_configuration/)                                               |
 | nkGlobalParameters.explicitUidInStatefulOperators  | Low        | boolean  | true                   | Should consistent [operator uids](https://ci.apache.org/projects/flink/flink-docs-stable/docs/ops/upgrading/#matching-operator-state) be used                                 |
 | nkGlobalParameters.useTypingResultTypeInformation  | Low        | boolean  | true                   | Enables using Nussknacker additional typing information for state serialization. It makes serialization much faster, currently consider it as experimental                    |
-| eventTimeMetricSlideDuration                       | Low        | duration | 10 seconds             | We use sliding window reservoir sampling for some metrics, this configures slide length                                                                                       |
 | componentsGroupMapping                             | Low        | map      |                        | Override default grouping of basic components in toolbox panels. Component names are keys, while values are toolbox panels name (e.g. sources, enrichers etc.)                |
 | componentActions                                   | Low        | `list[{id: string, title: string, icon: string, url: option[string], supportedComponentTypes: option[List[string]]}]` | `[{id: "usages", title: "Usages of component", icon: "/assets/components/actions/usages.svg"}]` | Component's actions configuration which are displayed at list of components.  Fields `title`, `icon`, `url` can contain templates: `$componentId` nad `$componentName` which are replaced by component data. Parameter `url` isn't required because some of actions like `usages` are build-in FE actions. Url is required when you want to do redirect to another page. Param `supportedComponentTypes` means component's types which can support actions.  |
